# Приложения (code examples)/network-manager-crm-web-app/backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Mini CRM API
  description: REST API for Mini CRM Telegram Mini App
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WorkerReportResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        projectName: { type: string }
        workerId: { type: string, format: uuid }
        workerName: { type: string }
        weekStart: { type: string, format: date }
        weekEnd: { type: string, format: date }
        status: { type: string, enum: [draft, review, accepted, returned, overdue] }
        totalHours: { type: string }
        totalAmount: { type: string }
        managerComment: { type: string }
    RejectWorkerReportRequest:
      type: object
      properties:
        comment: { type: string }
...
    ToolResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        serialNumber: { type: string }
        toolType: { type: string }
        controlType: { type: string, enum: [calibration, usage_limit] }
        status: { type: string, enum: [available, assigned, overdue, written_off] }
...
paths:
...
  /projects/{id}/workers:
    post:
      tags: [projects]
      security: [{ bearerAuth: [] }]
      summary: Assign worker to project
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Assigned }
  /projects/{id}/workers/{workerId}:
    delete:
      tags: [projects]
      security: [{ bearerAuth: [] }]
      summary: Remove worker from project
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: workerId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Removed }
...
  /workers:
    get:
      tags: [workers]
      security: [{ bearerAuth: [] }]
      summary: List workers
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: specialization, in: query, schema: { type: string } }
      responses:
        '200': { description: Worker list }
    post:
      tags: [workers]
      security: [{ bearerAuth: [] }]
      summary: Register worker application
      responses:
        '201': { description: Created }
  /workers/{id}:
    get:
      tags: [workers]
      security: [{ bearerAuth: [] }]
      summary: Get worker card
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Worker details }
    put:
      tags: [workers]
      security: [{ bearerAuth: [] }]
      summary: Update worker profile
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Updated }
  /workers/{id}/approve:
    post:
      tags: [workers]
      security: [{ bearerAuth: [] }]
      summary: Approve worker
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Approved }
  /workers/{id}/reject:
    post:
      tags: [workers]
      security: [{ bearerAuth: [] }]
      summary: Reject worker
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Rejected }
...
  /reports/worker:
    get:
      tags: [reports]
      security: [{ bearerAuth: [] }]
      summary: List worker reports
      responses:
        '200': { description: Report list }
    post:
      tags: [reports]
      security: [{ bearerAuth: [] }]
      summary: Create worker report
      responses:
        '201': { description: Created }
  /reports/worker/{id}:
    get:
      tags: [reports]
      security: [{ bearerAuth: [] }]
      summary: Get worker report
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkerReportResponse' }
    put:
      tags: [reports]
      security: [{ bearerAuth: [] }]
      summary: Update worker report
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Updated }
  /reports/worker/{id}/approve:
    post:
      tags: [reports]
      security: [{ bearerAuth: [] }]
      summary: Approve worker report
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Approved }
  /reports/worker/{id}/reject:
    post:
      tags: [reports]
      security: [{ bearerAuth: [] }]
      summary: Return worker report for correction
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RejectWorkerReportRequest' }
      responses:
        '200':
          description: Returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkerReportResponse' }
...
  /finance/workers:
    get:
      tags: [finance]
      security: [{ bearerAuth: [] }]
      summary: Finance by workers
      responses:
        '200': { description: Worker finance list }
  /finance/workers/{id}:
    get:
      tags: [finance]
      security: [{ bearerAuth: [] }]
      summary: Worker finance details
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Details }
...
  /tools:
    get:
      tags: [tools]
      security: [{ bearerAuth: [] }]
      summary: List tools
      responses:
        '200': { description: Tool list }
    post:
      tags: [tools]
      security: [{ bearerAuth: [] }]
      summary: Create tool
      responses:
        '201': { description: Created }
  /tools/{id}:
    get:
      tags: [tools]
      security: [{ bearerAuth: [] }]
      summary: Get tool card
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ToolResponse' }
    put:
      tags: [tools]
      security: [{ bearerAuth: [] }]
      summary: Update tool
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Updated }
  /tools/{id}/assign:
    post:
      tags: [tools]
      security: [{ bearerAuth: [] }]
      summary: Assign tool to project
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Assigned }
  /tools/{id}/return:
    post:
      tags: [tools]
      security: [{ bearerAuth: [] }]
      summary: Return tool from project
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Returned }
  /tools/{id}/calibrate:
    post:
      tags: [tools]
      security: [{ bearerAuth: [] }]
      summary: Record tool calibration
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Calibrated }
...